Settings generating script
Based on darkstorm's one from Discord, but with a bit more customizability,
less code repetition, and much better error-handling.

Also, in BASH format. So should work fine on Macs and linux machines.
If only we could somehow make a batch file to execute this, without WSL and other
required things...